refactor source folder
